+++ example/src/client.js
@@ -1,7 +1,7 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import App from './components/App';
 
 const mountNode = document.getElementById('mount');
 
-React.initializeTouchEvents(true);
-React.render(<App />, mountNode);
+ReactDOM.render(<App />, mountNode);